local-sql-requirements.txt
# Local SQL Execution Server Requirements

flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
